Fixed OpenMP issue when loading morphologies in parallel
Private vectors of geometry (Spheres, Cylinders and Cones) used to load individual morphologies where not defined in the right scope, meaning that they were accessed by several thread at the same time. Since std::vectors are not threadsafe, this was causing an exception. The purpose of this fix is to declare those private vectors in the right scope to make sure that they are accessed by one single thread.