Allow runtime changes of camera type.
Making some functions in RenderingParameters static. Refactoring the
type to string conversions for renderer and camera type to allow
making the conversion without an instance of RenderingParameters.

Adding CameraType::perspective and RendererType::basic as enum values
on their own, otherwise they become unreachable when a plugin to become
default is loaded.