Fix segfault when using Viewer and browser UI (#444)
When the browser is connecting to Brayns it changes the viewport size in preRender(). Since the Viewer is unaware of this change then the Viewer will try to draw the resized buffer with an incorrect viewport size. If the buffer is shrunk then this can lead to a segfault. Now instead, the actual width and height of the buffer is stored when rendering and then used when the Viewer is drawing.