Fix race condition causing deformed snapshots when using UI. (#477)
The problem was that the 'aspect' property in the camera was updated concurrently by both the snapshot renderer and the UI. This sometimes made the snapshot image stretched or compressed if the UI and the snapshot were using different aspect ratios. Now all the camera properties are cloned instead of shared when taking snapshots which fixes the issue.