Fix sporadic segfault (#513)
Sometimes the asynchronous simulation data is not loaded in time causing a segfault since the size of the data is set but not the actual data. Now the size is read from the actual data which should make these inconsistencies disappear. The same is now also done for transfer functions.