Immediate notify changes on the websocket
This fixes several issues regarding updates of objects that are not modified
by a rockets client, e.g. updates from plugins or programmatic updates like
the animation update while playback. It simplifies the postRender() logic a
lot, and also fixes #342.

To avoid excessive update 'spam', throttling is used on the change callbacks.