First iteration on material handling refactoring.
Added Scene::visitModels to apply a functor to all scene models
thread-safely.

brayns::Material now inherits from ProperpyObject. No special properties
have been defined except for the type. Two types are allowed: "default"
and "simulation". _castSimulationData member and accessors removed.

Added AbstractSimulationHandler::bind|unbind(MaterialPtr). The purpose of
this method is to attach simulation buffer data to the materials instead
of the renderer as it is now. For the moment, this method is used to change
the material types from "default" to "simulation" when a simulation
handler is assigned to a model.

OSPray specific Materials must now be registered to each renderer.
New logic ensures that the renderer is known at the momment the materials
are committed. Creating of ISPC object materials had to be deferred to a
later point. OSPRrayModel::commitMaterials has been added for this purpose.

OSPRayRenderer::createOSPRenderer made private.

The logic to obtain the simulation buffer objects now goes through the
material.

The default material is renamed from ExtendedObjMaterial to DefaultMaterial.
A new material, SimulationMaterial, has been added for the basic
and advanced simulation renderers. This material overrides the default
one and takes care of understanding the "simulation" material type to enable
to logic to get simulation buffer offsets from geometries. In practice,
the "cast_simulation_data" parameter has been renamed to "apply_simulation",
and it has been made dependent on the material type.

Refactorings in shader code.