Plugin loading and initialization split in two phases.
This is one step towards being able to extend command line parsing
with plugin specific options. As a requirement to it we need to load
plugins before parsing takes place.

Plugin handling has been moved to a separate class called PluginManager.

Registration of notification callbacks moved to the implementation in
the Rockets plugin.