More circuit and morphology loader refactorings and simplifications.
Loading morphologies through brain::Circuit, this enables supports
for morphology caching.

Removed the geometry parameters for placing morphologies in a grid.
This can be easily done with a Python script instead.

When displaying a circuit with spherical somas only, the soma radius
is now taken from statistics based on the mtypes. The radius multiplier
is not considered anymore (before the multiplier was the chosen radius,
which seemed bogus).