Fix crash with RocketsPlugin and VRPNPlugin loaded
Ensure camera is not adjusted by manipulator with empty scene bounds. As this
happened, the `to_json` for the camera returned a malformed string and the
exception thrown was not catched.

Also avoid broadcasting updates if there is no http client connected.

Fix also DeflectPlugin crash that would occur if the camera manipulator is
changed during runtime.